# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Business Development Agent (BDA) multi-tenant SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /sandbox-tests:
    get:
      operationId: listSandboxTests
      tags: [sandbox]
      summary: List sandbox test history
      responses:
        "200":
          description: Sandbox tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SandboxTest"
    post:
      operationId: runSandboxTest
      tags: [sandbox]
      summary: Run a sandbox test scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxTestInput"
      responses:
        "201":
          description: Sandbox test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxTest"
  /sandbox-tests/{id}/feedback:
    post:
      operationId: saveSandboxFeedback
      tags: [sandbox]
      summary: Save feedback and regenerate the sandbox test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxFeedbackInput"
      responses:
        "200":
          description: Regenerated sandbox test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxTest"
...
  /widget-settings:
    get:
      operationId: getWidgetSettings
      tags: [widget]
      summary: Get widget settings
      responses:
        "200":
          description: Widget settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetSettings"
    put:
      operationId: saveWidgetSettings
      tags: [widget]
      summary: Save widget settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetSettingsInput"
      responses:
        "200":
          description: Saved widget settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetSettings"
  /widget/config:
    get:
      operationId: getWidgetConfig
      tags: [widget]
      summary: Public widget config by client id
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public widget config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetConfig"
  /widget/interact:
    post:
      operationId: widgetInteract
      tags: [widget]
      summary: Public widget interaction that creates a lead and estimate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetInteractInput"
      responses:
        "200":
          description: AI response and generated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetInteractResult"
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        basePrice:
          type: ["number", "null"]
        hourlyRate:
          type: ["number", "null"]
        minimumPrice:
          type: ["number", "null"]
        estimatedDuration:
          type: ["string", "null"]
        active:
          type: boolean
    ServiceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        basePrice:
          type: ["number", "null"]
        hourlyRate:
          type: ["number", "null"]
        minimumPrice:
          type: ["number", "null"]
        estimatedDuration:
          type: string
        active:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        basePrice:
          type: ["number", "null"]
        hourlyRate:
          type: ["number", "null"]
        minimumPrice:
          type: ["number", "null"]
        estimatedDuration:
          type: string
        active:
          type: boolean
...
    Estimate:
      type: object
      required: [customerSummary, assumptions, invoiceLineItems, subtotal, taxes, totalEstimate, confidenceScore, followUpQuestions]
      properties:
        customerSummary:
          type: string
        assumptions:
          type: array
          items:
            type: string
        recommendedPriceLow:
          type: ["number", "null"]
        recommendedPriceHigh:
          type: ["number", "null"]
        invoiceLineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        taxes:
          type: number
        totalEstimate:
          type: number
        confidenceScore:
          type: number
        followUpQuestions:
          type: array
          items:
            type: string
    InvoiceLineItem:
      type: object
      required: [description, total]
      properties:
        description:
          type: string
        quantity:
          type: ["number", "null"]
        unitPrice:
          type: ["number", "null"]
        total:
          type: number
    SandboxTest:
      type: object
      required: [id, prompt, agentResponse, createdAt]
      properties:
        id:
          type: integer
        scenario:
          type: ["string", "null"]
        prompt:
          type: string
        agentResponse:
          type: string
        estimate:
          oneOf:
            - $ref: "#/components/schemas/Estimate"
            - type: "null"
        rating:
          type: ["integer", "null"]
        feedbackNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    SandboxTestInput:
      type: object
      required: [prompt]
      properties:
        scenario:
          type: string
        prompt:
          type: string
          minLength: 1
    SandboxFeedbackInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        feedbackNotes:
          type: string
...
    WidgetSettings:
      type: object
      required: [id, greeting, primaryColor, position, enabled]
      properties:
        id:
          type: integer
        greeting:
          type: string
        primaryColor:
          type: string
        position:
          type: string
          description: bottom-right or bottom-left
        enabled:
          type: boolean
    WidgetSettingsInput:
      type: object
      properties:
        greeting:
          type: string
        primaryColor:
          type: string
        position:
          type: string
        enabled:
          type: boolean
    WidgetConfig:
      type: object
      required: [clientId, businessName, greeting, primaryColor, position, enabled]
      properties:
        clientId:
          type: string
        businessName:
          type: string
        greeting:
          type: string
        primaryColor:
          type: string
        position:
          type: string
        enabled:
          type: boolean
    WidgetInteractInput:
      type: object
      required: [clientId, name, projectDescription]
      properties:
        clientId:
          type: string
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        projectDescription:
          type: string
          minLength: 1
    WidgetInteractResult:
      type: object
      required: [leadId, message, estimate]
      properties:
        leadId:
          type: integer
        message:
          type: string
        estimate:
          $ref: "#/components/schemas/Estimate"
...