# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Business Development Agent (BDA) multi-tenant SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  basePrice?: number | null;
  /** @nullable */
  hourlyRate?: number | null;
  /** @nullable */
  minimumPrice?: number | null;
  /** @nullable */
  estimatedDuration?: string | null;
  active: boolean;
}